Add BaseActivity so drawer can be used in multiple activities
