+++ app/src/main/java/com/example/habitup/View/BaseActivity.java
@@ -1,0 +1,102 @@
+package com.example.habitup.View;
+
+import android.support.design.widget.NavigationView;
+import android.support.v4.widget.DrawerLayout;
+import android.support.v7.app.ActionBarDrawerToggle;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.support.v7.widget.Toolbar;
+import android.view.MenuItem;
+import android.view.View;
+import android.widget.FrameLayout;
+
+import com.example.habitup.R;
+
+public class BaseActivity extends AppCompatActivity {
+
+    public DrawerLayout drawerLayout;
+    public ActionBarDrawerToggle drawerToggle;
+    private Toolbar toolbar;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    public void setContentView(int layoutResID) {
+        DrawerLayout fullView = (DrawerLayout) getLayoutInflater().inflate(R.layout.drawer, null);
+        FrameLayout container = (FrameLayout) fullView.findViewById(R.id.frame);
+        getLayoutInflater().inflate(layoutResID, container, true);
+
+        super.setContentView(fullView);
+        toolbar = (Toolbar) findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+
+        initNavigationDrawer();
+    }
+
+    public void initNavigationDrawer() {
+
+        // Taken from https://www.learn2crack.com/2016/03/android-material-design-sliding-navigation-drawer.html
+        NavigationView navigationView = (NavigationView)findViewById(R.id.navigation_view);
+        navigationView.setNavigationItemSelectedListener(new NavigationView.OnNavigationItemSelectedListener() {
+            @Override
+            public boolean onNavigationItemSelected(MenuItem menuItem) {
+
+                int id = menuItem.getItemId();
+
+                switch (id){
+                    case R.id.profile:
+                        drawerLayout.closeDrawers();
+                        break;
+                    case R.id.habits:
+                        drawerLayout.closeDrawers();
+                        break;
+                    case R.id.events:
+                        drawerLayout.closeDrawers();
+                        break;
+                    case R.id.friends:
+                        drawerLayout.closeDrawers();
+                        break;
+                    case R.id.requests:
+                        drawerLayout.closeDrawers();
+                        break;
+                    case R.id.finds:
+                        drawerLayout.closeDrawers();
+                        break;
+                    case R.id.map:
+                        drawerLayout.closeDrawers();
+                        break;
+                    case R.id.logout:
+                        drawerLayout.closeDrawers();
+                        break;
+                }
+                return true;
+            }
+        });
+
+        drawerLayout = (DrawerLayout) findViewById(R.id.drawer);
+        View header = navigationView.getHeaderView(0);
+
+        // Highlight profile list item
+        navigationView.getMenu().getItem(0).setChecked(true);
+
+        ActionBarDrawerToggle actionBarDrawerToggle = new ActionBarDrawerToggle(this,drawerLayout,toolbar,R.string.drawer_open,R.string.drawer_close){
+
+            @Override
+            public void onDrawerClosed(View v){
+                super.onDrawerClosed(v);
+            }
+
+            @Override
+            public void onDrawerOpened(View v) {
+                super.onDrawerOpened(v);
+            }
+        };
+
+        drawerLayout.addDrawerListener(actionBarDrawerToggle);
+        actionBarDrawerToggle.syncState();
+    }
+
+}
